fix: expose row grid geometry through engine types
